hazpost-backend/requirements.txt
flask==3.0.3
flask-limiter==3.8.0
apscheduler==3.10.4
requests==2.32.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
scikit-learn==1.5.1
numpy==1.26.4
gunicorn==22.0.0
